Add urlparse module
